backend/requirements.txt
alembic==1.0.8
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
Flask-Restless==0.17.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
mimerender==0.6.0
pprint==0.1
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-mimeparse==1.6.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.15.0
